requirements.txt
# Core Earth Engine and Geospatial (Docker optimized)
earthengine-api>=0.1.380
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-cloud-storage>=2.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# Geospatial libraries (using system GDAL)
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Machine Learning
scikit-learn>=1.3.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0

# HTTP and API
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0

# Task Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Performance Optimization
numba>=0.58.0
bottleneck>=1.3.0

# Date and Time
python-dateutil>=2.8.0
pytz>=2023.3

# Development and Testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Monitoring and Logging
psutil>=5.9.0

# Excel support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Additional for Docker
gunicorn>=21.2.0
